Refactor project schema and update homepage query for project references
